LayoutTests/imported/w3c/web-platform-tests/fetch/api/redirect/redirect-empty-location.any.worker-expected.txt


FAIL redirect response with empty Location, follow mode assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL redirect response with empty Location, manual mode assert_equals: expected "opaqueredirect" but got "basic"
